Add IIR filter sine wave generator
Sine wave generation is now much more efficient, not only
because we no longer call sin() for every sample, but because
each of those calls was itself calling out to JS due to our
having turned on the JS_MATH setting in Emscripten.

JS_MATH was originally turned on because doing so was tested
last year to be somewhat more performant on a per-sin() basis
on certain ancient browsers, even with the added overhead.
However, said performance gains could not be replicated when
re-investigating the effects of JS_MATH while testing this
commit. One wonders if they really existed at all, or if they
were some strange fluke.

Technically the sine wave generator still does have to call
sin() occasionally, but it ships with its own stripped-down
version of the same Sun math library sin() routine that
Emscripten provides by way of musl, reducing the code size
hit in the compiled .wasm from around 6 KiB to 2 KiB.